python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spindle-cli"
version = "0.1.0"
description = "A CLI tool for installing and managing reusable source code components with npm-like script running"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
keywords = ["cli", "package-manager", "components", "scripts"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "tomli>=2.0;python_version<'3.11'",
    "requests>=2.28",
]

[project.scripts]
spindle = "spindle.cli:main"
sp = "spindle.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]
